first steps to extract FigureManager* and friends from pyplot
This commit refactors backend_bases into two parts, the Gcf
independent parts, which moved to backends/_backend_bases.py, and the
Gcf dependent stuff which stayed in backend_bases.py.  The contents of
backends/_backend_bases.py are bulk imported into backend_bases so
this change should be backwards compatible.

Added an attribute to FigureManagerBase for the _key_press_handler
function.  This might be a stop-gap if the logic in `Gcf.destroy` should
be moved into the FigureManager or canvas classes.