CI: remove osx from build matrix
Travis has changed their free-to-open-source policy and these builds
are currently not running.

Only comment them out for now to un-break the build while we sort out
the situation.

Currently we are only running one OSX job on travis which is testing
on exteremely old versions of OSX (xcode9 on 10.12) which is not
supported on azure.  Losing this is not ideal, but I do not think
existential.