ENH: make error messages from exceeding agg cell block count clearer
The error messages now provide better instruction on how to fix the
problem.

There are some paths we can not split (one with hatching, fills, or
that claim they can not be simplified) so tell user that.

If we could have otherwise split up the path if the chunksize is set
to less than 100, which disables the feature, ask user to make it
bigger or make the path simplification more aggressive.

If the chunksize is bigger than the data, ask user to make it smaller
or make the path simplification more aggressive.

If the chunksize is smaller than the data, but still too big ask user
to make it smaller or make the path simplification more aggressive.

closes #19325