Fix decenter of image in gallery thumbnails
When using the default padding and the pydata-sphinx-theme, the image
is shifted away from the center. While that should be fixed in
sphinx-gallery (and I will report it there), this serves as a workaround
on the matplotlib side, because I don't think they will have it fixed
ready for the 3.5 release.